♻️  Use aggregate in saga event handlers
