+++ packages/applications/bootstrap/src/setupLauréat.ts
@@ -64,7 +64,7 @@ export const setupLauréat = async ({ sendEmail }: SetupLauréatDependencies) =>
   GarantiesFinancières.GarantiesFinancièresSaga.register();
   GarantiesFinancières.TypeGarantiesFinancièresSaga.register();
   Lauréat.LauréatSaga.register();
-  ReprésentantLégal.ReprésentantLégalSaga.register();
+  ReprésentantLégal.ReprésentantLégalSaga.register(loadAggregate);
   Actionnaire.ActionnaireSaga.register();
 
   const unsubscribeLauréatProjector = await subscribe<LauréatProjector.SubscriptionEvent>({
