+++ packages/applications/projectors/src/subscribers/lauréat/représentant-légal/handleChangementReprésentantLégalSupprimé.ts
@@ -2,12 +2,8 @@ import { ReprésentantLégal } from '@potentiel-domain/laureat';
 import { findProjection } from '@potentiel-infrastructure/pg-projections';
 import { Option } from '@potentiel-libraries/monads';
 import { getLogger } from '@potentiel-libraries/monitoring';
-import { fileExists, upload } from '@potentiel-libraries/file-storage';
-import { DocumentProjet } from '@potentiel-domain/document';
 
 import { removeProjection } from '../../../infrastructure';
-
-import { getSensitiveDocReplacement } from './getSensitiveDocReplacement';
 
 export const handleChangementReprésentantLégalSupprimé = async (
   event: ReprésentantLégal.ChangementReprésentantLégalSuppriméEvent,
@@ -29,20 +25,4 @@ export const handleChangementReprésentantLégalSupprimé = async (
   }
 
   await removeProjection(`changement-représentant-légal|${identifiantProjet}`);
-
-  const pièceJustificative = DocumentProjet.convertirEnValueType(
-    identifiantProjet,
-    ReprésentantLégal.TypeDocumentChangementReprésentantLégal.pièceJustificative.formatter(),
-    changementReprésentantLégal.demande.demandéLe,
-    changementReprésentantLégal.demande.pièceJustificative.format,
-  );
-
-  if (await fileExists(pièceJustificative.formatter())) {
-    await upload(
-      pièceJustificative.formatter(),
-      await getSensitiveDocReplacement(
-        'Document sensible supprimé automatiquement après suppression de la demande',
-      ),
-    );
-  }
 };
