--- packages/domain/lauréat/src/représentantLégal/saga/handleLauréatNotifié.ts
@@ -1,18 +1,0 @@
-import { mediator } from 'mediateur';
-
-import { DateTime, Email, IdentifiantProjet } from '@potentiel-domain/common';
-
-import { LauréatNotifiéEvent } from '../../lauréat';
-import { ImporterReprésentantLégalCommand } from '../importer/importerReprésentantLégal.command';
-
-export const handleLauréatNotifié = async ({
-  payload: { identifiantProjet, notifiéLe },
-}: LauréatNotifiéEvent) =>
-  mediator.send<ImporterReprésentantLégalCommand>({
-    type: 'Lauréat.ReprésentantLégal.Command.ImporterReprésentantLégal',
-    data: {
-      identifiantProjet: IdentifiantProjet.convertirEnValueType(identifiantProjet),
-      importéLe: DateTime.convertirEnValueType(notifiéLe),
-      importéPar: Email.system(),
-    },
-  });
