+++ packages/domain/lauréat/src/représentantLégal/saga/index.ts
@@ -1,14 +1,15 @@
 import { Message, MessageHandler, mediator } from 'mediateur';
 import { match } from 'ts-pattern';
 
+import { LoadAggregate } from '@potentiel-domain/core';
 import { TâchePlanifiéeExecutéeEvent } from '@potentiel-domain/tache-planifiee';
 
 import { LauréatNotifiéEvent } from '../../lauréat';
 import { AbandonAccordéEvent } from '../../abandon';
 
-import { handleLauréatNotifié } from './handleLauréatNotifié';
-import { handleTâchePlanifiéeGestionAutomatiqueDemandeChangementExecutée } from './handleTâchePlanifiéeGestionAutomatiqueDemandeChangementExecutée';
-import { handleAbandonAccordé } from './handleAbandonAccordé';
+import { buildLauréatNotifiéEventHandler } from './lauréatNotifié.eventHandler';
+import { buildTâchePlanifiéeGestionAutomatiqueDemandeChangementExecutéeEventHandler } from './tâchePlanifiéeGestionAutomatiqueDemandeChangementExecutée.eventHandler';
+import { buildAbandonAccordéEventHandler } from './abandonAccordé.eventHandler';
 
 export type SubscriptionEvent =
   | LauréatNotifiéEvent
@@ -17,7 +18,12 @@ export type SubscriptionEvent =
 
 export type Execute = Message<'System.Lauréat.ReprésentantLégal.Saga.Execute', SubscriptionEvent>;
 
-export const register = () => {
+export const register = (loadAggregate: LoadAggregate) => {
+  const handleAbandonAccordé = buildAbandonAccordéEventHandler(loadAggregate);
+  const handleLauréatNotifié = buildLauréatNotifiéEventHandler();
+  const handleTâchePlanifiéeGestionAutomatiqueDemandeChangementExecutée =
+    buildTâchePlanifiéeGestionAutomatiqueDemandeChangementExecutéeEventHandler(loadAggregate);
+
   const handler: MessageHandler<Execute> = async (event) =>
     match(event)
       .with({ type: 'LauréatNotifié-V1' }, handleLauréatNotifié)
